+++ stack/source/6lowpan/ws/ws_pae_controller.h
@@ -682,7 +682,7 @@ int8_t ws_pae_controller_auth_cb_register(protocol_interface_info_entry_t *inter
  * \param ticks elapsed ticks
  *
  */
-void ws_pae_controller_fast_timer(uint16_t ticks);
+void ws_pae_controller_fast_timer(int ticks);
 
 /**
  * ws_pae_controller_slow_timer PAE controller slow timer call
@@ -690,7 +690,7 @@ void ws_pae_controller_fast_timer(uint16_t ticks);
  * \param seconds elapsed seconds
  *
  */
-void ws_pae_controller_slow_timer(uint16_t seconds);
+void ws_pae_controller_slow_timer(int seconds);
 
 struct nvm_tlv *ws_pae_controller_nvm_tlv_get(protocol_interface_info_entry_t *interface_ptr);
